Final schedule and plan

+++ backend/backend/src/main/java/be/backend/repository/MachineRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface MachineRepository extends JpaRepository<Machine,Integer> {
-    List<Machine> findByLineIdAndStatus(Long lineId, String status);
+        List<Machine> findByLineIdAndStatus(Integer lineId, String status);
 
-}
+    }
